+++ src/main/java/io/github/ImpactDevelopment/installer/setting/settings/InstallationModeSetting.java
@@ -19,7 +19,7 @@ package io.github.ImpactDevelopment.installer.setting.settings;
 
 import io.github.ImpactDevelopment.installer.setting.ChoiceSetting;
 import io.github.ImpactDevelopment.installer.setting.InstallationConfig;
-import io.github.ImpactDevelopment.installer.versions.InstallationModeOptions;
+import io.github.ImpactDevelopment.installer.target.InstallationModeOptions;
 
 import java.util.Arrays;
 import java.util.List;
