+++ src/main/java/io/github/ImpactDevelopment/installer/utils/Fetcher.java
@@ -15,10 +15,28 @@
  * along with Impact Installer.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.github.ImpactDevelopment.installer.versions;
+package io.github.ImpactDevelopment.installer.utils;
+
+import org.apache.commons.io.IOUtils;
 
 import java.io.IOException;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
 
-public interface InstallationMode {
-    void apply() throws IOException;  // not gonna lie this is me when i enter sicko mode
+/**
+ * Put all the URL fetching in one place so that it can be logged
+ */
+public class Fetcher {
+    public static String fetch(String url) {
+        return new String(fetchBytes(url), StandardCharsets.UTF_8);
+    }
+
+    public static byte[] fetchBytes(String url) {
+        System.out.println("DOWNLOADING " + url);
+        try {
+            return IOUtils.toByteArray(new URL(url).openStream());
+        } catch (IOException e) {
+            throw new RuntimeException("Unable to fetch " + url, e);
+        }
+    }
 }
